% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-mean-n-min.R
\name{calc_mean_n_min}
\alias{calc_mean_n_min}
\title{Calculate mean function addressing item-level missing data using proration}
\usage{
calc_mean_n_min(data, n_min)
}
\arguments{
\item{data}{Dataframe}

\item{n_min}{Minimum number of scores needed to calculate mean}
}
\description{
Calculate mean function addressing item-level missing data using proration
}
