% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sum.R
\name{calc_sum}
\alias{calc_sum}
\title{Calculate sum of variables}
\usage{
calc_sum(data, id_str, var_str, session_str, n_min, item_scores = FALSE,
  sort_sum_item = TRUE, sep = "_")
}
\arguments{
\item{data}{Wide dataframe.}

\item{id_str}{String of identifier variable.}

\item{var_str}{String of variable to calculate sum for.}

\item{session_str}{String of session number.}

\item{n_min}{Minimum number of available scores to calculate sum.}

\item{item_scores}{Add item scores after sum.}

\item{sort_sum_item}{Logical, if TRUE and multiple sessions then output dataframe will be organised sum_timepoint followed by all items for that timepoint, if FALSE all sums will come after id variable followed by all items.}

\item{sep}{seperator for variable names.}
}
\description{
Calculate sum addressing item-level missing data using proration
}
