% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-range.R
\name{check_item_range}
\alias{check_item_range}
\title{Check range of item scores}
\usage{
check_item_range(data, id_str, var_str, session_str, range_min, range_max)
}
\arguments{
\item{data}{Wide dataframe.}

\item{id_str}{String of identifier (id) variable.}

\item{var_str}{String of variable to check.}

\item{session_str}{String of session number.}

\item{range_min}{Number of minimum possible value of scale.}

\item{range_max}{Number of maximum possible value of scale.}
}
\description{
Function to check if all item scores are within the possible range. If values are out of range, show the scores for that case together with id.
}
