% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-var-names.R
\name{create_var_names}
\alias{create_var_names}
\title{Create variable names}
\usage{
create_var_names(var_str, sess_start, sess_end, sess_str = "s",
  var_sess_sep = "_")
}
\arguments{
\item{var_str}{String for variable}

\item{sess_start}{First session}

\item{sess_end}{Last session}

\item{sess_str}{String, abbreviation for timepoint, default s for session}

\item{var_sess_sep}{String, seperator between var_str and sess_str}
}
\description{
Create variable names for repeated measures
}
