% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-item-vars.R
\name{rename_item_vars}
\alias{rename_item_vars}
\title{Rename item variables}
\usage{
rename_item_vars(data, id_str, data_session, new_session = data_session,
  data_str_measure, prefix_data_str_session,
  new_str_measure = data_str_measure, prefix_new_str_session = "s",
  prefix_new_str_item = "i")
}
\arguments{
\item{data}{Dataset}

\item{id_str}{String, id variable}

\item{data_str_measure}{Sting, unique string identifying measure}

\item{prefix_data_str_session}{String pattern in data that identifies session/timepoint}

\item{new_str_measure}{String, if unhappy with name of measure in data, you can change it here}

\item{prefix_new_str_session}{String, idicator for session/timepoint, default "s"}

\item{prefix_new_str_item}{String, idicator for item, default "i"}

\item{session}{Numeric or String identifying session/timepoint}
}
\description{
Rename item variables consistently, this works only for variables named InMeasure1 which stands for Item 1 at Intake Assessment for Measure, or Session1Measure5 wich stands for Item 5 at Session 1 of Measure
}
